refactor(codegen): remove control handlers
