+++ codegen/src/visitor/mod.rs
@@ -11,7 +11,6 @@ use wasmparser::{for_each_operator, BlockType, BrTable, Ieee32, Ieee64, MemArg, 
 
 mod call;
 mod control;
-mod handlers;
 mod local;
 mod log;
